introduced atlas module
